+++ .gitignore
@@ -1,0 +1,21 @@
+*.iml
+.gradle
+/local.properties
+/captures
+*.ipr
+*.iws
+.idea/
+.classpath
+.project
+bin/
+gen/
+out/
+target/
+local.properties
+classes/
+tmp/
+project.properties
+.DS_Store
+.settings
+build/
+jacoco.exec
