+++ .travis.yml
@@ -53,7 +53,7 @@ script:
   - chmod 755 gradlew
   - ls -la
   #Run jacocoTestReport -> Runs all unit tests & instrumentation tests
-  - travis_wait ./gradlew :opensrp-chw-familyPlanning:jacocoTestReport coveralls --stacktrace
+  - travis_wait ./gradlew :opensrp-chw-fp:jacocoTestReport coveralls --stacktrace
 
 notifications:
   slack:
