+++ opensrp-chw-family-planning/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="org.smartregister.familyPlanning" />
+<manifest package="org.smartregister.fp" />
