--- opensrp-chw-family-planning/src/main/java/org/smartregister/chw/familyPlanning/FamilyPlanningLibrary.java
@@ -1,110 +1,0 @@
-package org.smartregister.chw.familyPlanning;
-
-import org.smartregister.Context;
-import org.smartregister.CoreLibrary;
-import org.smartregister.repository.Repository;
-import org.smartregister.repository.UniqueIdRepository;
-import org.smartregister.sync.ClientProcessorForJava;
-import org.smartregister.sync.helper.ECSyncHelper;
-
-import id.zelory.compressor.Compressor;
-
-public class FamilyPlanningLibrary {
-    private static FamilyPlanningLibrary instance;
-
-    private final Context context;
-    private final Repository repository;
-
-    private int applicationVersion;
-    private int databaseVersion;
-
-    private UniqueIdRepository uniqueIdRepository;
-    private ECSyncHelper syncHelper;
-
-    private ClientProcessorForJava clientProcessorForJava;
-    private Compressor compressor;
-
-    public static void init(Context context, Repository repository, int applicationVersion, int databaseVersion) {
-        if (instance == null) {
-            instance = new FamilyPlanningLibrary(context, repository, applicationVersion, databaseVersion);
-        }
-    }
-
-    public static FamilyPlanningLibrary getInstance() {
-        if (instance == null) {
-            throw new IllegalStateException(" Instance does not exist!!! Call "
-                    + CoreLibrary.class.getName()
-                    + ".init method in the onCreate method of "
-                    + "your Application class ");
-        }
-        return instance;
-    }
-
-    private FamilyPlanningLibrary(Context contextArg, Repository repositoryArg, int applicationVersion, int databaseVersion) {
-        this.context = contextArg;
-        this.repository = repositoryArg;
-        this.applicationVersion = applicationVersion;
-        this.databaseVersion = databaseVersion;
-    }
-
-    public Context context() {
-        return context;
-    }
-
-    public Repository getRepository() {
-        return repository;
-    }
-
-    public int getApplicationVersion() {
-        return applicationVersion;
-    }
-
-    public int getDatabaseVersion() {
-        return databaseVersion;
-    }
-
-    public UniqueIdRepository getUniqueIdRepository() {
-        if (uniqueIdRepository == null) {
-            uniqueIdRepository = new UniqueIdRepository(getRepository());
-        }
-        return uniqueIdRepository;
-    }
-
-    public ECSyncHelper getEcSyncHelper() {
-        if (syncHelper == null) {
-            syncHelper = ECSyncHelper.getInstance(context().applicationContext());
-        }
-        return syncHelper;
-    }
-
-
-    public ClientProcessorForJava getClientProcessorForJava() {
-        if (clientProcessorForJava == null) {
-            clientProcessorForJava = ClientProcessorForJava.getInstance(context().applicationContext());
-        }
-        return clientProcessorForJava;
-    }
-
-    public void setClientProcessorForJava(ClientProcessorForJava clientProcessorForJava) {
-        this.clientProcessorForJava = clientProcessorForJava;
-    }
-
-    public Compressor getCompressor() {
-        if (compressor == null) {
-            compressor = Compressor.getDefault(context().applicationContext());
-        }
-        return compressor;
-    }
-
-    /**
-     * Use this method when testing.
-     * It should replace org.smartregister.Context#setInstance(org.smartregister.Context, org.smartregister.repository.Repository) which has been removed
-     *
-     * @param context
-     */
-    public static void reset(Context context, Repository repository, int applicationVersion, int databaseVersion) {
-        if (context != null) {
-            instance = new FamilyPlanningLibrary(context, repository, applicationVersion, databaseVersion);
-        }
-    }
-}
