--- opensrp-chw-family-planning/src/main/java/org/smartregister/chw/familyPlanning/contract/BaseFamilyPlanningRegisterFragmentContract.java
@@ -1,53 +1,0 @@
-package org.smartregister.chw.familyPlanning.contract;
-
-import org.json.JSONArray;
-import org.smartregister.configurableviews.model.Field;
-import org.smartregister.configurableviews.model.RegisterConfiguration;
-import org.smartregister.configurableviews.model.ViewConfiguration;
-import org.smartregister.domain.Response;
-import org.smartregister.view.contract.BaseRegisterFragmentContract;
-
-import java.util.List;
-import java.util.Set;
-
-public interface BaseFamilyPlanningRegisterFragmentContract {
-
-
-    interface View extends BaseRegisterFragmentContract.View {
-
-        void initializeAdapter(Set<org.smartregister.configurableviews.model.View> visibleColumns);
-
-        Presenter presenter();
-
-    }
-
-    interface Presenter extends BaseRegisterFragmentContract.Presenter {
-
-        void updateSortAndFilter(List<Field> filterList, Field sortField);
-
-        String getMainCondition();
-
-        String getDefaultSortQuery();
-
-    }
-
-    interface Model {
-
-        RegisterConfiguration defaultRegisterConfiguration();
-
-        ViewConfiguration getViewConfiguration(String viewConfigurationIdentifier);
-
-        Set<org.smartregister.configurableviews.model.View> getRegisterActiveColumns(String viewConfigurationIdentifier);
-
-        String countSelect(String tableName, String mainCondition);
-
-        String mainSelect(String tableName, String mainCondition);
-
-        String getFilterText(List<Field> filterList, String filter);
-
-        String getSortText(Field sortField);
-
-        JSONArray getJsonArray(Response<String> response);
-
-    }
-}
