--- opensrp-chw-family-planning/src/main/java/org/smartregister/chw/familyPlanning/model/BaseFamilyPlanningRegisterModel.java
@@ -1,41 +1,0 @@
-package org.smartregister.chw.familyPlanning.model;
-
-import org.json.JSONObject;
-import org.smartregister.chw.familyPlanning.contract.BaseFamilyPlanningRegisterContract;
-
-import java.util.List;
-
-public class BaseFamilyPlanningRegisterModel implements BaseFamilyPlanningRegisterContract.Model  {
-    @Override
-    public void registerViewConfigurations(List<String> viewIdentifiers) {
-//        implement
-
-    }
-
-    @Override
-    public void unregisterViewConfiguration(List<String> viewIdentifiers) {
-//        implement
-
-    }
-
-    @Override
-    public void saveLanguage(String language) {
-//        implement
-
-    }
-
-    @Override
-    public String getLocationId(String locationName) {
-        return null;
-    }
-
-    @Override
-    public JSONObject getFormAsJson(String formName, String entityId, String currentLocationId) throws Exception {
-        return null;
-    }
-
-    @Override
-    public String getInitials() {
-        return null;
-    }
-}
