--- opensrp-chw-family-planning/src/main/java/org/smartregister/chw/familyPlanning/presenter/BaseFamilyPlanningRegisterFragmentPresenter.java
@@ -1,98 +1,0 @@
-package org.smartregister.chw.familyPlanning.presenter;
-
-import org.apache.commons.lang3.StringUtils;
-import org.smartregister.chw.familyPlanning.contract.BaseFamilyPlanningRegisterFragmentContract;
-import org.smartregister.chw.familyPlanning.util.DBConstants;
-import org.smartregister.configurableviews.model.Field;
-import org.smartregister.configurableviews.model.RegisterConfiguration;
-import org.smartregister.configurableviews.model.View;
-import org.smartregister.configurableviews.model.ViewConfiguration;
-
-import java.lang.ref.WeakReference;
-import java.util.List;
-import java.util.Set;
-import java.util.TreeSet;
-
-public class BaseFamilyPlanningRegisterFragmentPresenter implements BaseFamilyPlanningRegisterFragmentContract.Presenter {
-
-    protected WeakReference<BaseFamilyPlanningRegisterFragmentContract.View> viewReference;
-
-    protected BaseFamilyPlanningRegisterFragmentContract.Model model;
-
-    protected RegisterConfiguration config;
-
-    protected Set<View> visibleColumns = new TreeSet<>();
-    protected String viewConfigurationIdentifier;
-
-    public BaseFamilyPlanningRegisterFragmentPresenter(BaseFamilyPlanningRegisterFragmentContract.View view, BaseFamilyPlanningRegisterFragmentContract.Model model, String viewConfigurationIdentifier) {
-        this.viewReference = new WeakReference<>(view);
-        this.model = model;
-        this.viewConfigurationIdentifier = viewConfigurationIdentifier;
-        this.config = model.defaultRegisterConfiguration();
-    }
-
-    @Override
-    public void updateSortAndFilter(List<Field> filterList, Field sortField) {
-//        implement
-    }
-
-    @Override
-    public String getMainCondition() {
-        return "";
-    }
-
-    @Override
-    public String getDefaultSortQuery() {
-        return "";
-    }
-
-    @Override
-    public void processViewConfigurations() {
-        if (StringUtils.isBlank(viewConfigurationIdentifier)) {
-            return;
-        }
-
-        ViewConfiguration viewConfiguration = model.getViewConfiguration(viewConfigurationIdentifier);
-        if (viewConfiguration != null) {
-            config = (RegisterConfiguration) viewConfiguration.getMetadata();
-            this.visibleColumns = model.getRegisterActiveColumns(viewConfigurationIdentifier);
-        }
-
-        if (config.getSearchBarText() != null && getView() != null) {
-            getView().updateSearchBarHint(config.getSearchBarText());
-        }
-    }
-
-    @Override
-    public void initializeQueries(String mainCondition) {
-        String tableName = DBConstants.TABLE_NAME.FAMILY_PLANNING_TABLE;
-
-        String countSelect = model.countSelect(tableName, mainCondition);
-        String mainSelect = model.mainSelect(tableName, mainCondition);
-
-        getView().initializeQueryParams(tableName, countSelect, mainSelect);
-        getView().initializeAdapter(visibleColumns);
-
-        getView().countExecute();
-        getView().filterandSortInInitializeQueries();
-    }
-
-    protected BaseFamilyPlanningRegisterFragmentContract.View getView() {
-        if (viewReference != null)
-            return viewReference.get();
-        else
-            return null;
-    }
-
-    @Override
-    public void startSync() {
-//        implement
-
-    }
-
-    @Override
-    public void searchGlobally(String s) {
-//        implement
-
-    }
-}
