--- opensrp-chw-family-planning/src/main/java/org/smartregister/chw/familyPlanning/util/ConfigHelper.java
@@ -1,33 +1,0 @@
-package org.smartregister.chw.familyPlanning.util;
-
-import android.content.Context;
-
-import org.smartregister.configurableviews.model.Field;
-import org.smartregister.configurableviews.model.RegisterConfiguration;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class ConfigHelper {
-    public static RegisterConfiguration defaultRegisterConfiguration(Context context) {
-        if (context == null) {
-            return null;
-        }
-
-        RegisterConfiguration config = new RegisterConfiguration();
-        config.setEnableAdvancedSearch(false);
-        config.setEnableFilterList(false);
-        config.setEnableSortList(false);
-        config.setSearchBarText(context.getString(org.smartregister.R.string.search_hint));
-        config.setEnableJsonViews(false);
-
-        List<Field> filers = new ArrayList<>();
-        config.setFilterFields(filers);
-
-        List<Field> sortFields = new ArrayList<>();
-        config.setSortFields(sortFields);
-
-        return config;
-    }
-
-}
