--- opensrp-chw-family-planning/src/main/java/org/smartregister/chw/familyPlanning/util/Constants.java
@@ -1,20 +1,0 @@
-package org.smartregister.chw.familyPlanning.util;
-
-public interface Constants {
-
-    int REQUEST_CODE_GET_JSON = 2244;
-    String ENCOUNTER_TYPE = "encounter_type";
-
-    interface JSON_FORM_EXTRA {
-        String JSON = "json";
-        String NEXT = "next";
-    }
-
-    interface EVENT_TYPE {
-        String FAMILY_PLANNING_REGISTRATION = "Family Planning Registration";
-    }
-
-    interface FORMS {
-        String FAMILY_PLANNING_REGISTRATION_FORM = "family_planning_registration";
-    }
-}
