+++ opensrp-chw-family-planning/src/main/java/org/smartregister/chw/fp/provider/BaseFpRegisterProvider.java
@@ -1,0 +1,123 @@
+package org.smartregister.chw.fp.provider;
+
+import android.content.Context;
+import android.database.Cursor;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+import android.widget.TextView;
+
+import org.smartregister.cursoradapter.RecyclerViewProvider;
+import org.smartregister.fp.R;
+import org.smartregister.view.contract.SmartRegisterClient;
+import org.smartregister.view.contract.SmartRegisterClients;
+import org.smartregister.view.dialog.FilterOption;
+import org.smartregister.view.dialog.ServiceModeOption;
+import org.smartregister.view.dialog.SortOption;
+import org.smartregister.view.viewholder.OnClickFormLauncher;
+
+import java.text.MessageFormat;
+
+public class BaseFpRegisterProvider implements RecyclerViewProvider<BaseFpRegisterProvider.RegisterViewHolder> {
+
+    private final LayoutInflater inflater;
+
+    private View.OnClickListener paginationClickListener;
+
+    private Context context;
+
+    public BaseFpRegisterProvider(Context context, View.OnClickListener paginationClickListener) {
+//        TODO add onClickListener and commonRepository to constructor
+
+        inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+        this.paginationClickListener = paginationClickListener;
+
+        this.context = context;
+    }
+
+    @Override
+    public void getView(Cursor cursor, SmartRegisterClient smartRegisterClient, RegisterViewHolder registerViewHolder) {
+//        checkout how family module is implemented
+    }
+
+    @Override
+    public void getFooterView(RecyclerView.ViewHolder viewHolder, int currentPageCount, int totalPageCount, boolean hasNext, boolean hasPrevious) {
+        FooterViewHolder footerViewHolder = (FooterViewHolder) viewHolder;
+        footerViewHolder.pageInfoView.setText(
+                MessageFormat.format(context.getString(org.smartregister.R.string.str_page_info), currentPageCount,
+                        totalPageCount));
+
+        footerViewHolder.nextPageView.setVisibility(hasNext ? View.VISIBLE : View.INVISIBLE);
+        footerViewHolder.previousPageView.setVisibility(hasPrevious ? View.VISIBLE : View.INVISIBLE);
+
+        footerViewHolder.nextPageView.setOnClickListener(paginationClickListener);
+        footerViewHolder.previousPageView.setOnClickListener(paginationClickListener);
+    }
+
+    @Override
+    public SmartRegisterClients updateClients(FilterOption filterOption, ServiceModeOption serviceModeOption, FilterOption filterOption1, SortOption sortOption) {
+        return null;
+    }
+
+    @Override
+    public void onServiceModeSelected(ServiceModeOption serviceModeOption) {
+//        implement
+
+    }
+
+    @Override
+    public OnClickFormLauncher newFormLauncher(String s, String s1, String s2) {
+        return null;
+    }
+
+    @Override
+    public LayoutInflater inflater() {
+        return inflater;
+    }
+
+    @Override
+    public RegisterViewHolder createViewHolder(ViewGroup parent) {
+        View view = inflater.inflate(R.layout.family_planning_register_list_row, parent, false);
+        return new RegisterViewHolder(view);
+    }
+
+    @Override
+    public RecyclerView.ViewHolder createFooterHolder(ViewGroup parent) {
+        View view = inflater.inflate(R.layout.smart_register_pagination, parent, false);
+        return new FooterViewHolder(view);
+    }
+
+    @Override
+    public boolean isFooterViewHolder(RecyclerView.ViewHolder viewHolder) {
+        return FooterViewHolder.class.isInstance(viewHolder);
+    }
+
+    // implement place holder view
+    public class RegisterViewHolder extends RecyclerView.ViewHolder {
+
+        public RegisterViewHolder(View itemView) {
+            super(itemView);
+
+
+        }
+    }
+
+    // implement footer view
+    public class FooterViewHolder extends RecyclerView.ViewHolder {
+        public TextView pageInfoView;
+        public Button nextPageView;
+        public Button previousPageView;
+
+        public FooterViewHolder(View view) {
+            super(view);
+
+            nextPageView = view.findViewById(org.smartregister.R.id.btn_next_page);
+            previousPageView = view.findViewById(org.smartregister.R.id.btn_previous_page);
+            pageInfoView = view.findViewById(org.smartregister.R.id.txt_page_info);
+        }
+    }
+
+
+}
