+++ opensrp-chw-family-planning/src/main/java/org/smartregister/chw/fp/util/DBConstants.java
@@ -1,0 +1,44 @@
+package org.smartregister.chw.fp.util;
+
+/**
+ * Created by keyman on 13/11/2018.
+ */
+public class DBConstants {
+    public final class TABLE_NAME {
+        public static final String FAMILY_PLANNING_TABLE = "ec_family_planning";
+    }
+
+    public final class KEY {
+        public static final String ID = "_ID";
+        public static final String FIRST_NAME = "first_name";
+        public static final String MIDDLE_NAME = "middle_name";
+        public static final String LAST_NAME = "last_name";
+        public static final String BASE_ENTITY_ID = "base_entity_id";
+        public static final String RELATIONAL_ID = "relational_id";
+        public static final String DOB = "dob";//Date Of Birth
+        public static final String DOD = "dod";
+        public static final String GENDER = "gender";
+        public static final String UNIQUE_ID = "unique_id";
+        public static final String LAST_INTERACTED_WITH = "last_interacted_with";
+        public static final String CONTACT_STATUS = "contact_status";
+        public static final String VILLAGE_TOWN = "village_town";
+        public static final String FAMILY_HEAD = "family_head";
+        public static final String PRIMARY_CAREGIVER = "primary_caregiver";
+        public static final String QUATER_CLAN = "quarter_clan";
+        public static final String STREET = "street";
+        public static final String LANDMARK = "landmark";
+        public static final String GPS = "gps";
+        public static final String FAMILY_SOURCE_INCOME = "fam_source_income";
+        public static final String PHONE_NUMBER = "phone_number";
+        public static final String OTHER_PHONE_NUMBER = "other_phone_number";
+        public static final String HIGHEST_EDU_LEVEL = "highest_edu_level";
+        public static final String DATE_REMOVED = "date_removed";
+        public static final String DATE_LAST_HOME_VISIT = "last_home_visit";
+        public static final String DATE_VISIT_NOT_DONE = "visit_not_done";
+
+        public static final String OBJECT_ID = "object_id";
+        public static final String OBJECT_RELATIONAL_ID = "object_relational_id";
+
+        public static final String ENTITY_TYPE = "entity_type";
+    }
+}
