+++ sample/src/main/AndroidManifest.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="org.smartregister.familyPlanning">
+    package="org.smartregister.fp">
 
     <application
-        android:name="org.smartregister.familyPlanning.application.SampleApplication"
+        android:name="org.smartregister.fp.application.SampleApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="org.smartregister.chw.familyPlanning.activity.FamilyPlanningRegisterActivity"
+            android:name="org.smartregister.chw.fp.activity.FpRegisterActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
