--- sample/src/main/java/org/smartregister/chw/familyPlanning/activity/FamilyPlanningRegisterActivity.java
@@ -1,21 +1,0 @@
-package org.smartregister.chw.familyPlanning.activity;
-
-import android.view.MenuItem;
-
-public class FamilyPlanningRegisterActivity extends BaseFamilyPlanningRegisterActivity {
-
-    @Override
-    protected void registerBottomNavigation() {
-        super.registerBottomNavigation();
-
-        if(bottomNavigationView != null){
-            MenuItem clients = bottomNavigationView.getMenu().findItem(org.smartregister.R.id.action_clients);
-            if (clients != null) {
-                clients.setTitle("");
-            }
-
-            bottomNavigationView.getMenu().removeItem(org.smartregister.R.id.action_search);
-            bottomNavigationView.getMenu().removeItem(org.smartregister.R.id.action_library);
-        }
-    }
-}
