+++ sample/src/main/java/org/smartregister/chw/fp/activity/FpRegisterActivity.java
@@ -1,0 +1,21 @@
+package org.smartregister.chw.fp.activity;
+
+import android.view.MenuItem;
+
+public class FpRegisterActivity extends BaseFpRegisterActivity {
+
+    @Override
+    protected void registerBottomNavigation() {
+        super.registerBottomNavigation();
+
+        if(bottomNavigationView != null){
+            MenuItem clients = bottomNavigationView.getMenu().findItem(org.smartregister.R.id.action_clients);
+            if (clients != null) {
+                clients.setTitle("");
+            }
+
+            bottomNavigationView.getMenu().removeItem(org.smartregister.R.id.action_search);
+            bottomNavigationView.getMenu().removeItem(org.smartregister.R.id.action_library);
+        }
+    }
+}
