--- sample/src/main/java/org/smartregister/familyPlanning/repository/SampleRepository.java
@@ -1,116 +1,0 @@
-package org.smartregister.familyPlanning.repository;
-
-import android.content.Context;
-import android.util.Log;
-
-import net.sqlcipher.database.SQLiteDatabase;
-
-import org.smartregister.AllConstants;
-import org.smartregister.configurableviews.repository.ConfigurableViewsRepository;
-import org.smartregister.familyPlanning.BuildConfig;
-import org.smartregister.familyPlanning.application.SampleApplication;
-import org.smartregister.repository.EventClientRepository;
-import org.smartregister.repository.Repository;
-import org.smartregister.repository.SettingsRepository;
-import org.smartregister.repository.UniqueIdRepository;
-
-/**
- * Created by keyman on 28/07/2017.
- */
-public class SampleRepository extends Repository {
-
-    private static final String TAG = SampleRepository.class.getCanonicalName();
-    protected SQLiteDatabase readableDatabase;
-    protected SQLiteDatabase writableDatabase;
-    private Context context;
-    private String password = "Sample_PASS";
-
-    public SampleRepository(Context context, org.smartregister.Context openSRPContext) {
-        super(context, AllConstants.DATABASE_NAME, BuildConfig.DATABASE_VERSION, openSRPContext.session(), SampleApplication.createCommonFtsObject(), openSRPContext.sharedRepositoriesArray());
-        this.context = context;
-    }
-
-    @Override
-    public void onCreate(SQLiteDatabase database) {
-        super.onCreate(database);
-        EventClientRepository.createTable(database, EventClientRepository.Table.client, EventClientRepository.client_column.values());
-        EventClientRepository.createTable(database, EventClientRepository.Table.event, EventClientRepository.event_column.values());
-
-        ConfigurableViewsRepository.createTable(database);
-
-        UniqueIdRepository.createTable(database);
-        SettingsRepository.onUpgrade(database);
-
-        onUpgrade(database, 1, BuildConfig.DATABASE_VERSION);
-    }
-
-    @Override
-    public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        Log.w(SampleRepository.class.getName(),
-                "Upgrading database from version " + oldVersion + " to "
-                        + newVersion + ", which will destroy all old data");
-
-        int upgradeTo = oldVersion + 1;
-        while (upgradeTo <= newVersion) {
-            switch (upgradeTo) {
-                case 2:
-                    //upgradeToVersion2(db);
-                    break;
-                default:
-                    break;
-            }
-            upgradeTo++;
-        }
-    }
-
-
-    @Override
-    public SQLiteDatabase getReadableDatabase() {
-        return getReadableDatabase(password);
-    }
-
-    @Override
-    public SQLiteDatabase getWritableDatabase() {
-        return getWritableDatabase(password);
-    }
-
-    @Override
-    public synchronized SQLiteDatabase getReadableDatabase(String password) {
-        try {
-            if (readableDatabase == null || !readableDatabase.isOpen()) {
-                if (readableDatabase != null) {
-                    readableDatabase.close();
-                }
-                readableDatabase = super.getReadableDatabase(password);
-            }
-            return readableDatabase;
-        } catch (Exception e) {
-            Log.e(TAG, "Database Error. " + e.getMessage());
-            return null;
-        }
-
-    }
-
-    @Override
-    public synchronized SQLiteDatabase getWritableDatabase(String password) {
-        if (writableDatabase == null || !writableDatabase.isOpen()) {
-            if (writableDatabase != null) {
-                writableDatabase.close();
-            }
-            writableDatabase = super.getWritableDatabase(password);
-        }
-        return writableDatabase;
-    }
-
-    @Override
-    public synchronized void close() {
-        if (readableDatabase != null) {
-            readableDatabase.close();
-        }
-
-        if (writableDatabase != null) {
-            writableDatabase.close();
-        }
-        super.close();
-    }
-}
